Log.Backend.StandardOutput

Description

Stdout logging back-end.

Documentation

withSimpleStdOutLogger :: MonadUnliftIO m => (Logger -> m r) -> m r

Deprecated: Use withStdOutLogger instead

withStdOutLogger :: MonadUnliftIO m => (Logger -> m r) -> m r

Create a logger that prints messages to standard output for the duration of the given action.

withJsonStdOutLogger :: MonadUnliftIO m => (Logger -> m r) -> m r

Create a logger that prints messages in the JSON format to standard output for the duration of the given action.