Log.Backend.Text

Description

A logger that produces in-memory Text values. Mainly useful for testing.

Documentation

withSimpleTextLogger :: MonadUnliftIO m => (Logger -> m r) -> m (Text, r)

Create an in-memory logger for the duration of the given action, returning both the result of the action and the logger's output as a Text value afterwards.