Log.Class

Description

The MonadLog type class of monads with logging capabilities.

Documentation

data UTCTime

This is the simplest representation of UTC. It consists of the day number, and a time offset from midnight. Note that if a day has a leap second added to it, it will have 86401 seconds.

class Monad m => MonadLog m where

Represents the family of monads with logging capabilities. Each MonadLog carries with it some associated state (the logging environment) that can be modified locally with localData and localDomain.

Methods

logMessage

Arguments

:: LogLevel
Log level.

-> Text
Log message.

-> Value
Additional data associated with the message.

-> m ()

Write a message to the log.

localData :: [Pair] -> m a -> m a

Extend the additional data associated with each log message locally.

localDomain :: Text -> m a -> m a

Extend the current application domain locally.

localMaxLogLevel :: LogLevel -> m a -> m a

Override the current maximum log level.

getLoggerEnv :: m LoggerEnv

Get current LoggerEnv object. Useful for construction of logging functions that work in a different monad, see getLoggerIO as an example.

logAttention :: (MonadLog m, ToJSON a) => Text -> a -> m ()

Log a message and its associated data using current time as the event time and the LogAttention log level.

logInfo :: (MonadLog m, ToJSON a) => Text -> a -> m ()

Log a message and its associated data using current time as the event time and the LogInfo log level.

logTrace :: (MonadLog m, ToJSON a) => Text -> a -> m ()

Log a message and its associated data using current time as the event time and the LogTrace log level.

logAttention_ :: MonadLog m => Text -> m ()

Like logAttention, but without any additional associated data.

logInfo_ :: MonadLog m => Text -> m ()

Like logInfo, but without any additional associated data.

logTrace_ :: MonadLog m => Text -> m ()

Like logTrace, but without any additional associated data.