Log.Data

Description

Basic data types used throughout the package.

Documentation

data LogLevel

Available log levels. Note that ordering in this definintion determines what the maximum log level is. See leMaxLogLevel.

Constructors

LogAttention

LogInfo

LogTrace

showLogLevel :: LogLevel -> Text

readLogLevel :: Text -> LogLevel

This function is partial.

readLogLevelEither :: Text -> Either String LogLevel

data LogMessage

Represents message to be logged.

Constructors

LogMessage

Fields

lmComponent :: !Text

Component of an application.

lmDomain :: ![Text]

Application log domain.

lmTime :: !UTCTime

Time of the logged event.

lmLevel :: !LogLevel

Log level.

lmMessage :: !Text

Message to be logged.

lmData :: !Value

Additional data associated with the message.

showLogMessage

Arguments

:: Maybe UTCTime
The time that message was added to the log.

-> LogMessage
The actual message.

-> Text

Render a LogMessage to Text.

defaultLogLevel :: LogLevel

The default log level. Returns LogInfo.