Log.Monad

Description

The LogT monad transformer for adding logging capabilities to any monad.

Documentation

data Logger

An object used for communication with a logger thread that outputs LogMessages using e.g. PostgreSQL, Elasticsearch or stdout (depending on the back-end chosen).

data LoggerEnv

The state that every LogT carries around.

Constructors

LoggerEnv

Fields

leLogger :: !Logger

The Logger to use.

leComponent :: !Text

Current application component.

leDomain :: ![Text]

Current application domain.

leData :: ![Pair]

Additional data to be merged with the log message's data.

leMaxLogLevel :: LogLevel

The maximum log level allowed to be logged.

type InnerLogT = ReaderT LoggerEnv

newtype LogT m a

Monad transformer that adds logging capabilities to the underlying monad.

Constructors

LogT

Fields

unLogT :: InnerLogT m a

runLogT

Arguments

:: Text
Application component name to use.

-> Logger
The logging back-end to use.

-> LogLevel
The maximum log level allowed to be logged. Only messages less or equal than this level with be logged.

-> LogT m a
The LogT computation to run.

-> m a

Run a LogT computation.

Note that in the case of asynchronous/bulk loggers runLogT doesn't guarantee that all messages are actually written to the log once it finishes. Use withPGLogger or withElasticSearchLogger for that.

mapLogT :: (m a -> n b) -> LogT m a -> LogT n b

Transform the computation inside a LogT.

logMessageIO :: LoggerEnv -> UTCTime -> LogLevel -> Text -> Value -> IO ()

Base implementation of logMessage for use with a specific LoggerEnv. Useful for reimplementation of MonadLog instance.

getLoggerIO :: MonadLog m => m (UTCTime -> LogLevel -> Text -> Value -> IO ())

Return an IO action that logs messages using the current MonadLog context. Useful for interfacing with libraries such as aws or amazonka that accept logging callbacks operating in IO.